dockerize-llms/falcon7b/falcon7b/server/requirements.txt
-i https://pypi.org/simple

argparse==1.4.0
aiocontextvars==0.2.2
cython==0.29.23
msgpack-numpy==0.4.8
msgpack==1.0.2
python-json-logger==2.0.2
pyyaml==6.0.0
fastapi==0.95.0
uvicorn==0.21.1
uvloop==0.17.0
psutil==5.9.4
joblib==1.2.0
requests==2.31.0
